Allow GIF decoding up to the GIF width limit
Change GIF decoder line widths and row counters to use int, size the
decoder row buffer for the maximum GIF dimension, and pass the real GIF
image width through gif_view. Add coverage for small, signed-limit, and
maximum-width GIF rows.

+++ libid/engine/engine_timer.cpp
@@ -70,7 +70,7 @@ static int timer(const TimerType type, int (*fn)(), ...)
         break;
     case TimerType::DECODER:
         i = va_arg(arg_marker, int);
-        out = static_cast<int>(decoder(static_cast<short>(i)));
+        out = decoder(i);
         break;
     case TimerType::ENCODER:
         out = encoder();
