+++ libid/include/io/decoder.h
@@ -9,7 +9,7 @@ namespace id::io
 
 extern int                 (*g_out_line)(Byte *, int);
 
-short decoder(short line_width);
+int decoder(int line_width);
 void set_byte_buff(Byte *ptr);
 
 } // namespace id::io
